Updates, refactoring; ready for CR.
