--- src/main/java/io/anserini/index/generator/LuceneHtmlGenerator.java
@@ -1,10 +1,0 @@
-package io.anserini.index.generator;
-
-import io.anserini.document.SourceDocument;
-import io.anserini.index.transform.LuceneHtmlParserTransform;
-
-public class LuceneHtmlGenerator extends LuceneDocumentGenerator<SourceDocument> {
-  public LuceneHtmlGenerator() {
-    super(new LuceneHtmlParserTransform());
-  }
-}
