--- src/main/java/io/anserini/index/transform/StringTransformation.java
@@ -1,9 +1,0 @@
-package io.anserini.index.transform;
-
-import java.util.function.UnaryOperator;
-
-/**
- * Created by jimmylin on 3/25/17.
- */
-public abstract class StringTransformation implements UnaryOperator<String> {
-}
